requirements.txt
moviepy~=1.0.3
Pillow~=9.5.0
pytube~=15.0.0